+++ manager-infra/src/main/java/kr/co/aim/infra/persistence/springdatajpa/ProcessStatusHistoryJpaRepository.java
@@ -3,5 +3,5 @@ package kr.co.aim.infra.persistence.springdatajpa;
 import kr.co.aim.infra.persistence.entity.ProcessStatusHistoryEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProcessStatusHistoryJpaRepository extends JpaRepository<ProcessStatusHistoryEntity, Integer> {
+public interface ProcessStatusHistoryJpaRepository extends JpaRepository<ProcessStatusHistoryEntity, Long> {
 }
