+++ manager-infra/src/main/java/kr/co/aim/infra/persistence/springdatajpa/PurgeConfigJpaRepository.java
@@ -3,5 +3,5 @@ package kr.co.aim.infra.persistence.springdatajpa;
 import kr.co.aim.infra.persistence.entity.PurgeConfigEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PurgeConfigJpaRepository extends JpaRepository<PurgeConfigEntity, Integer> {
+public interface PurgeConfigJpaRepository extends JpaRepository<PurgeConfigEntity, Long> {
 }
